+++ Frends.AzureBlobStorage.WriteBlob/Frends.AzureBlobStorage.WriteBlob/Definitions/Input.cs
@@ -1,4 +1,4 @@
-﻿using System.ComponentModel;
+using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
 using Frends.AzureBlobStorage.WriteBlob.Attributes;
 using Frends.AzureBlobStorage.WriteBlob.Enums;
@@ -6,10 +6,55 @@ using Frends.AzureBlobStorage.WriteBlob.Enums;
 namespace Frends.AzureBlobStorage.WriteBlob.Definitions;
 
 /// <summary>
-/// Destination parameters.
+/// Input parameters.
 /// </summary>
-public class Destination
+public class Input
 {
+    /// <summary>
+    /// Selection of source types.
+    /// </summary>
+    /// <example>SourceType.String</example>
+    [DefaultValue(SourceType.Bytes)]
+    public SourceType SourceType { get; set; }
+
+    /// <summary>
+    /// Source content in string format.
+    /// </summary>
+    /// <example>This is test content</example>
+    [UIHint(nameof(SourceType), "", SourceType.String)]
+    public string ContentString { get; set; }
+
+    /// <summary>
+    /// Source content in byte array.
+    /// </summary>
+    /// <example>VGhpcyBpcyB0ZXN0</example>
+    [DisplayFormat(DataFormatString = "Expression")]
+    [UIHint(nameof(SourceType), "", SourceType.Bytes)]
+    public byte[] ContentBytes { get; set; }
+
+    /// <summary>
+    /// Set desired content-encoding.
+    /// Defaults to UTF8 BOM.
+    /// </summary>
+    /// <example>utf8</example>
+    [DefaultValue(FileEncoding.UTF8)]
+    public FileEncoding Encoding { get; set; }
+
+    /// <summary>
+    /// Enables BOM for UTF-8.
+    /// </summary>
+    /// <example>true</example>
+    [UIHint(nameof(Encoding), "", FileEncoding.UTF8)]
+    [DefaultValue(true)]
+    public bool EnableBOM { get; set; }
+
+    /// <summary>
+    /// Content encoding as string. A partial list of possible encodings: https://en.wikipedia.org/wiki/Windows_code_page#List.
+    /// </summary>
+    /// <example>windows-1252</example>
+    [UIHint(nameof(Encoding), "", FileEncoding.Other)]
+    public string FileEncodingString { get; set; }
+
     /// <summary>
     /// Name of the Azure Blob Storage container.
     /// Task will convert all letters to lowercase.
@@ -24,7 +69,7 @@ public class Destination
     /// Name of the blob. Blob name can also be folder structure and folders will be created to Blob Storage.
     /// See more info: https://learn.microsoft.com/en-us/rest/api/storageservices/naming-and-referencing-containers--blobs--and-metadata#blob-names
     /// </summary>
-    /// <example>BlobName.txt; C:\folder\blobName.txt</example>
+    /// <example>myblob.txt</example>
     [DisplayFormat(DataFormatString = "Text")]
     [NotEmptyString]
     public string BlobName { get; set; }
@@ -54,7 +99,7 @@ public class Destination
     public HandleExistingFile HandleExistingFile { get; set; }
 
     /// <summary>
-    /// Should the string be compressed before sending?
+    /// Should the content be compressed before sending?
     /// </summary>
     /// <example>false</example>
     [DefaultValue(false)]